extract_to_storage/requirements.txt
requests==2.31.0
google.cloud==0.34.0
google-cloud-storage==2.10.0